refactor: todo-library
todo-service провайдится в todo-module, создается enum filter
