+++ projects/todo-library/src/lib/todo/types/todo.ts
@@ -3,3 +3,9 @@ export interface Todo {
   title: string;
   isCompleted: boolean;
 }
+
+export enum Filter {
+  All = 1,
+  Active,
+  Completed
+}
